feat: Implementações e refatorações de serviço de produtos e exceções personalizadas - Adicionadas exceções personalizadas para serviços de produto, melhorando o tratamento de erros e a clareza das respostas para casos específicos. - Refatorado código de ProdutoServiceImpl e EstoqueServiceImpl:   - Simplificação de métodos com redundâncias.   - Aplicação de Optional para evitar NullPointerException.   - Ajuste na conversão de DTOs com ProdutoMapper para maior consistência. - Melhorias gerais de código:   - Padronização de nomenclatura de variáveis e métodos.   - Remoção de comentários obsoletos.   - Ajustes em mensagens de log para facilitar debugging. - Preparação para futuras implementações de validação de regras de negócio e integração com gamificação. - Implementação na OrdemDeProducaoService da automação de toda a lógica de cálculo e resposta para ordens de corte, incluindo geração automática de cortes e retalhos conforme exemplo de payload fornecido, além de correção dos erros de duplicidade e robustez do código.

+++ src/main/resources/db/migration/V5__Refactor_to_hybrid_production_model.sql .sql
@@ -25,8 +25,8 @@ CREATE TABLE ordens_de_corte (
                                  quantidade_produzida INT NOT NULL,
                                  modo_calculo VARCHAR(50) NOT NULL,
     -- Dimensões Finais (seja calculado ou manual)
-                                 largura_final_cm NUMERIC(10, 2) NOT NULL,
-                                 comprimento_final_cm NUMERIC(10, 2) NOT NULL,
+                                 largura_final_cm DOUBLE PRECISION NOT NULL,
+                                 comprimento_final_cm DOUBLE PRECISION NOT NULL,
     -- Campos de auditoria
                                  data_criacao TIMESTAMP WITH TIME ZONE NOT NULL,
                                  motivo VARCHAR(255),
@@ -34,4 +34,3 @@ CREATE TABLE ordens_de_corte (
                                  CONSTRAINT fk_ordem_produto FOREIGN KEY (produto_id) REFERENCES produtos(id),
                                  CONSTRAINT fk_ordem_lote FOREIGN KEY (lote_principal_id) REFERENCES lotes_materia_prima(id)
 );
-
